src/producer_lambda_image/requirements.txt
boto3==1.28.51
botocore==1.31.51
alpaca-py==0.42.1